Make it possible to create a VNFCInstance and other improvements
Now it is possible to scale out using the cli.
The cli up to now was not able to process more than three arguments now it is.
Changed exit status from 0 to 1 if an error occures.
Improved the messages to the user in case of some exceptions.
Fixed the bug that methods which were named the same in the agent classes but had different parameters were not recognised correctly.